Tables formatting in ndy.md
Some tables have been reformatted for better look in commandline
editors. Though I've left the 'The list of parameters' table intact as
there's so much text it, at least for me, looks cleaner the way it was
originally - a formatted one will most likely not fit well even in a
terminal on a FHD display.

Signed-off-by: Złotowłosy Kocur <camaro96masakruje@yahoo.com>